refactor: move out common pieces from AuditConsumers into AbstractAuditConsumer, moves consumers from legacy to consumers package
